added sent, recv with size

+++ tcp-base/tcputils.h
@@ -3,7 +3,7 @@
 
 #include "tcp.h"
 
-int gt_send_size(int sockfd, tcp_packet_t *packet);
-int gt_recv_size(int sockfd, tcp_packet_t *packet);
+size_t gt_send_size(int sockfd, tcp_packet_t *packet);
+size_t gt_recv_size(int sockfd, tcp_packet_t *packet);
 
 #endif
